Remove legacy FxA sync flows (Fixes #9144)
