+++ bedrock/base/templates/base-protocol.html
@@ -122,6 +122,7 @@
       <!--[if !IE]><!-->
         {{ js_bundle('lib') }}
         {{ js_bundle('ui') }}
+        {{ js_bundle('fxa') }}
         {{ js_bundle('data') }}
       <!--<![endif]-->
     {% endblock %}
