+++ bedrock/careers/templates/careers/base.html
@@ -21,6 +21,7 @@
     <script src="{{ static('js/careers/libs/protocol-base.min.js') }}"></script>
     <script src="{{ static('js/careers/libs/protocol-utils.min.js') }}"></script>
     <script src="{{ static('js/careers/libs/protocol-supports.min.js') }}"></script>
+    {{ js_bundle('fxa') }}
     {{ js_bundle('data') }}
   <!--<![endif]-->
 {% endblock %}
