+++ bedrock/firefox/templates/firefox/browsers/index.html
@@ -166,6 +166,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('fxa_form') }}
   {{ js_bundle('firefox-browsers-products') }}
 
   {% if show_firefox_app_store_banner %}
