+++ bedrock/firefox/templates/firefox/firstrun/firstrun.html
@@ -72,5 +72,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_firstrun_quantum') }}
+  {{ js_bundle('fxa_form') }}
 {% endblock %}
