+++ bedrock/firefox/templates/firefox/nightly/whatsnew.html
@@ -58,7 +58,7 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_nightly_whatsnew') }}
+  {{ js_bundle('firefox_whatsnew_update') }}
 {% endblock %}
 
 {# Exclude stub attribution for in-product pages: issus 9620 #}
