+++ bedrock/firefox/templates/firefox/pocket.html
@@ -110,5 +110,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('product_pocket') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
