+++ bedrock/firefox/templates/firefox/privacy/products.html
@@ -205,5 +205,6 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('fxa_form') }}
   {{ js_bundle('firefox-privacy-products') }}
 {% endblock %}
