+++ bedrock/firefox/templates/firefox/products/index.html
@@ -160,6 +160,7 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('fxa_form') }}
   {{ js_bundle('firefox-browsers-products') }}
 
   {% if show_firefox_app_store_banner %}
