+++ bedrock/firefox/templates/firefox/sync.html
@@ -113,5 +113,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('sync') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
