+++ bedrock/firefox/templates/firefox/welcome/page1.html
@@ -108,5 +108,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_welcome_page') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
