+++ bedrock/firefox/templates/firefox/welcome/page2.html
@@ -86,5 +86,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_welcome_page') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
