+++ bedrock/firefox/templates/firefox/welcome/page3.html
@@ -92,5 +92,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_welcome_page') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
