+++ bedrock/firefox/templates/firefox/whatsnew/index-account.html
@@ -75,5 +75,6 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('fxa_product_button') }}
   {{ js_bundle('firefox_whatsnew_account') }}
 {% endblock %}
