+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx93-de.html
@@ -54,5 +54,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_whatsnew_93') }}
+  {{ js_bundle('firefox_whatsnew_update') }}
 {% endblock %}
