+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx93-fr.html
@@ -54,5 +54,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_whatsnew_93') }}
+  {{ js_bundle('firefox_whatsnew_update') }}
 {% endblock %}
