+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx93-nl.html
@@ -53,5 +53,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_whatsnew_93') }}
+  {{ js_bundle('firefox_whatsnew_update') }}
 {% endblock %}
