+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx95-de.html
@@ -75,5 +75,6 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_whatsnew_95_de') }}
+  {{ js_bundle('firefox_whatsnew_update') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
