+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx95-en.html
@@ -39,5 +39,5 @@
 {% endblock %}
 
 {% block js %}
-  {{ js_bundle('firefox_whatsnew_95_en') }}
+  {{ js_bundle('firefox_whatsnew_update') }}
 {% endblock %}
