+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx96-de.html
@@ -131,4 +131,5 @@
 
 {% block js %}
   {{ js_bundle('firefox_whatsnew_96_eu') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
