+++ bedrock/firefox/templates/firefox/whatsnew/whatsnew-fx96-fr.html
@@ -130,4 +130,5 @@
 
 {% block js %}
   {{ js_bundle('firefox_whatsnew_96_eu') }}
+  {{ js_bundle('fxa_product_button') }}
 {% endblock %}
