+++ bedrock/products/templates/products/vpn/landing.html
@@ -293,5 +293,6 @@
 {% endblock %}
 
 {% block js %}
+  {{ js_bundle('fxa_product_button') }}
   {{ js_bundle('mozilla-vpn-landing') }}
 {% endblock %}
