--- media/js/firefox/whatsnew/account-conditional-ctas.js
@@ -1,17 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-
-(function (Mozilla) {
-    'use strict';
-
-    Mozilla.Client.getFxaDetails(function (details) {
-        if (details.setup) {
-            Mozilla.FxaProductButton.init();
-        } else {
-            Mozilla.FxaForm.init();
-        }
-    });
-})(window.Mozilla);
