+++ media/js/products/vpn/landing.js
@@ -25,15 +25,6 @@
 (function () {
     'use strict';
 
-    // initiate FxA flow metrics after subscription URLs have been set.
-    if (typeof Mozilla.FxaProductButton !== 'undefined') {
-        Mozilla.FxaProductButton.init();
-    }
-})();
-
-(function () {
-    'use strict';
-
     function openFaqItem(id) {
         var faq = document.getElementById(id);
