+++ media/static-bundles.json
@@ -1065,13 +1065,6 @@
     },
     {
       "files": [
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js"
-      ],
-      "name": "firefox_welcome_page"
-    },
-    {
-      "files": [
         "protocol/js/protocol-modal.js",
         "js/base/send-to-device.js",
         "js/firefox/welcome/welcome4.js"
@@ -1101,31 +1094,9 @@
     },
     {
       "files": [
-        "js/firefox/whatsnew/up-to-date.js"
-      ],
-      "name": "firefox_whatsnew_93"
-    },
-    {
-      "files": [
-        "js/firefox/whatsnew/up-to-date.js",
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js"
-      ],
-      "name": "firefox_whatsnew_95_de"
-    },
-    {
-      "files": [
-        "js/firefox/whatsnew/up-to-date.js"
-      ],
-      "name": "firefox_whatsnew_95_en"
-    },
-    {
-      "files": [
         "js/firefox/whatsnew/up-to-date.js",
         "js/base/mozilla-fxa.js",
-        "js/base/mozilla-fxa-init.js",
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js"
+        "js/base/mozilla-fxa-init.js"
       ],
       "name": "firefox_whatsnew_96_eu"
     },
@@ -1333,8 +1304,6 @@
     },
     {
       "files": [
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js",
         "protocol/js/protocol-sticky-promo.js",
         "js/firefox/sticky-promo.js",
         "js/firefox/home/master.js"
@@ -1352,8 +1321,6 @@
       "files": [
         "js/base/mozilla-fxa.js",
         "js/base/mozilla-fxa-init.js",
-        "js/base/mozilla-fxa-form.js",
-        "js/base/mozilla-fxa-form-init.js",
         "protocol/js/protocol-sticky-promo.js",
         "js/firefox/sticky-promo.js",
         "js/firefox/browsers-products.js"
@@ -1397,18 +1364,21 @@
       "files": [
         "js/base/mozilla-fxa.js",
         "js/base/mozilla-fxa-init.js",
-        "js/base/mozilla-fxa-form.js",
-        "js/base/mozilla-fxa-product-button.js",
-        "js/firefox/accounts-2019.js"
+        "js/firefox/accounts-2019.es6.js"
       ],
       "name": "firefox_accounts_2019"
     },
     {
       "files": [
-        "js/base/mozilla-fxa-form.js",
-        "js/base/mozilla-fxa-form-init.js"
+        "js/base/fxa-form-init.es6.js"
       ],
-      "name": "firefox_firstrun_quantum"
+      "name": "fxa_form"
+    },
+    {
+      "files": [
+        "js/base/fxa-product-button-init.es6.js"
+      ],
+      "name": "fxa_product_button"
     },
     {
       "files": [
@@ -1435,17 +1405,9 @@
     },
     {
       "files": [
-        "js/firefox/whatsnew/up-to-date.js"
-      ],
-      "name": "firefox_nightly_whatsnew"
-    },
-    {
-      "files": [
         "js/firefox/whatsnew/up-to-date.js",
         "js/base/mozilla-fxa.js",
         "js/base/mozilla-fxa-init.js",
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js",
         "js/base/send-to-device.js",
         "js/firefox/whatsnew/whatsnew-account.js"
       ],
@@ -1453,8 +1415,6 @@
     },
     {
       "files": [
-        "js/base/mozilla-fxa-form.js",
-        "js/base/mozilla-fxa-form-init.js",
         "js/firefox/privacy/products.js"
       ],
       "name": "firefox-privacy-products"
@@ -1504,18 +1464,22 @@
         "protocol/js/protocol-details.js",
         "js/base/protocol/protocol-footer.js",
         "js/base/protocol/protocol-sub-navigation.js",
-        "js/base/mozilla-fxa-link.js",
-        "js/base/mozilla-fxa-link-init.js",
         "js/base/base-page-init.js"
       ],
       "name": "ui"
     },
     {
       "files": [
+        "js/base/fxa-link.es6.js",
+        "js/base/fxa-utm.es6.js",
+        "js/base/fxa-bundle-init.es6.js"
+      ],
+      "name": "fxa"
+    },
+    {
+      "files": [
         "js/base/core-datalayer.js",
-        "js/base/core-datalayer-init.js",
-        "js/base/fxa-utm-referral.js",
-        "js/base/fxa-utm-referral-init.js"
+        "js/base/core-datalayer-init.js"
       ],
       "name": "data"
     },
@@ -1547,13 +1511,6 @@
         "js/press/speaker-request.js"
       ],
       "name": "press_speaker_request"
-    },
-    {
-      "files": [
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js"
-      ],
-      "name": "product_pocket"
     },
     {
       "files": [
@@ -1615,14 +1572,6 @@
     },
     {
       "files": [
-        "js/base/mozilla-fxa-product-button.js",
-        "js/base/mozilla-fxa-product-button-init.js"
-      ],
-      "name": "sync"
-    },
-    {
-      "files": [
-        "js/base/mozilla-fxa-product-button.js",
         "js/products/vpn/landing.js"
       ],
       "name": "mozilla-vpn-landing"
