+++ tests/unit/karma.conf.js
@@ -27,32 +27,28 @@ module.exports = function (config) {
             'media/js/base/core-datalayer.js',
             'media/js/base/dnt-helper.js',
             'media/js/base/experiment-utils.es6.js',
-            'media/js/base/fxa-utm-referral.js',
             'media/js/base/mozilla-convert.js',
             'media/js/base/mozilla-fxa.js',
-            'media/js/base/mozilla-fxa-form.js',
-            'media/js/base/mozilla-fxa-link.js',
-            'media/js/base/mozilla-fxa-product-button.js',
             'media/js/base/mozilla-pixel.js',
             'media/js/base/mozilla-smoothscroll.js',
             'media/js/base/send-to-device.js',
             'media/js/base/stub-attribution.js',
             'media/js/firefox/all/all-downloads-unified.js',
             'media/js/firefox/new/common/thanks.js',
-            'tests/unit/spec/base/mozilla-banner.js',
-            'tests/unit/spec/base/mozilla-run.js',
             'tests/unit/spec/base/core-datalayer-page-id.js',
             'tests/unit/spec/base/core-datalayer.js',
             'tests/unit/spec/base/dnt-helper.js',
             'tests/unit/spec/base/experiment-utils.js',
-            'tests/unit/spec/base/fxa-utm-referral.js',
-            'tests/unit/spec/base/mozilla-convert.js',
+            'tests/unit/spec/base/fxa-form.js',
+            'tests/unit/spec/base/fxa-link.js',
+            'tests/unit/spec/base/fxa-utm.js',
+            'tests/unit/spec/base/mozilla-banner.js',
             'tests/unit/spec/base/mozilla-client.js',
+            'tests/unit/spec/base/mozilla-convert.js',
+            'tests/unit/spec/base/fxa-product-button.js',
             'tests/unit/spec/base/mozilla-fxa.js',
-            'tests/unit/spec/base/mozilla-fxa-form.js',
-            'tests/unit/spec/base/mozilla-fxa-link.js',
-            'tests/unit/spec/base/mozilla-fxa-product-button.js',
             'tests/unit/spec/base/mozilla-pixel.js',
+            'tests/unit/spec/base/mozilla-run.js',
             'tests/unit/spec/base/mozilla-smoothscroll.js',
             'tests/unit/spec/base/mozilla-utils.js',
             'tests/unit/spec/base/search-params.js',
